+++ tsfile/src/main/java/org/apache/iotdb/tsfile/read/common/block/column/IntColumn.java
@@ -32,7 +32,7 @@ import static org.apache.iotdb.tsfile.read.common.block.column.ColumnUtil.checkV
 
 public class IntColumn implements Column {
 
-  private static final int INSTANCE_SIZE = ClassLayout.parseClass(IntColumn.class).instanceSize();
+  private static final long INSTANCE_SIZE = ClassLayout.parseClass(IntColumn.class).instanceSize();
   public static final int SIZE_IN_BYTES_PER_POSITION = Integer.BYTES + Byte.BYTES;
 
   private final int arrayOffset;
